requirements.txt
fastapi
uvicorn
pinecone
langchain_openai
python-dotenv
langchain_text_splitters